requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.2
django-environ==0.13.0
django-redis>=5.4.0
eventlet>=0.36.0
idna==3.11
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
qrcode==8.2
redis==7.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
django-celery-results==2.6.0

